[public-api] Use standard cli setup
